Add conditional logic to script, fix order
Host is now checked to see if it's part of a swarm before trying to
remove stacks and services.

Logic added to see if there are actually any services/stacks/containers
before trying to pass them to the 'rm' commands.

Moved prune operation to after leaving the swarm so that swarm bridge
networks are cleaned up without having to re-run the script.

Basic argument parsing added to handle '--volumes' and '--all' being
passed to the script for additional pruning options.